packages/manager/pyproject.toml
[tool.poetry]
name = "manager"
version = "0.1.0"
description = ""
authors = ["mablin7 <magyarb0201@gmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "3.7.9"
PyYAML = "^5.4.1"
nanoid = "^2.0.0"
docker = "^5.0.0"
minio = "^7.1.0"
pyzmq = "^22.1.0"
api-types-py = {path = "../api_types_py", develop = true}

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
